Move the Nemotron mamba stub under the family package and document VLA export.
Keep Hub load stubs next to the Nemotron spec, and mention PI05/GR00T run_vla.py in the LLM tools README.
